requirements.txt
altair==4.1.0
argon2-cffi==21.1.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.0
bleach==4.1.0
blinker==1.4
Bottleneck==1.3.2
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.4
cycler==0.10.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fonttools==4.25.0
gitdb==4.0.7
GitPython==3.1.24
greenlet @ file:///D:/bld/greenlet_1632928344703/work
idna==3.2
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonschema==3.2.0
jupyter-client==7.0.5
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver @ file:///C:/ci/kiwisolver_1612282606037/work
MarkupSafe==2.0.1
matplotlib @ file:///C:/ci/matplotlib-suite_1628011511716/work
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
munkres==1.1.4
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.4
numexpr @ file:///C:/ci/numexpr_1618856738664/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1626271903025/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
packaging==21.0
pandas @ file:///C:/ci/pandas_1629682609359/work
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow @ file:///C:/ci/pillow_1602770972588/work
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
pyarrow==5.0.0
pycairo==1.20.1
pycparser==2.20
pydeck==0.7.0
Pygments==2.10.0
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
pyrsistent==0.18.0
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==301
pywinpty==1.1.4
pyzmq==22.3.0
reportlab==3.5.68
requests==2.26.0
scikit-learn @ file:///C:/ci/scikit-learn_1622739500535/work
scipy @ file:///C:/ci/scipy_1630606942584/work
Send2Trash==1.8.0
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==4.0.0
SQLAlchemy @ file:///D:/bld/sqlalchemy_1632383544096/work
streamlit==0.89.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
toml==0.10.2
toolz==0.11.1
tornado @ file:///C:/ci/tornado_1606942392901/work
traitlets==5.1.0
typing-extensions==3.10.0.2
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.7
validators==0.18.2
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
